requirements.txt
numpy==2.4.2
openai==2.24.0
pandas==1.5.3
python-dotenv==1.2.1
Requests==2.32.5
streamlit==1.54.0
